Created new component for password reset
